+++ crates/holochain/src/conductor/conductor.rs
@@ -68,7 +68,7 @@ use std::sync::atomic::AtomicBool;
 use std::sync::Arc;
 use tracing::*;
 
-#[cfg(any(test, feature = "test_utils"))]
+#[cfg(feature = "test_utils")]
 use super::handle::MockConductorHandleT;
 
 mod share;
@@ -1548,8 +1548,7 @@ mod builder {
                 p2p_batch_senders: Arc::new(parking_lot::Mutex::new(HashMap::new())),
                 p2p_metrics_env: envs.p2p_metrics(),
                 db_sync_level: self.config.db_sync_level,
-                #[cfg(any(test, feature = "test_utils"))]
-                dev_settings: parking_lot::RwLock::new(DevSettings::default()),
+                dev_settings: Default::default(),
             });
 
             // Install extra DNAs, in particular: