+++ crates/kitsune_p2p/kitsune_p2p/src/lib.rs
@@ -19,10 +19,11 @@ pub use config::*;
 mod spawn;
 pub use spawn::*;
 
-#[cfg(any(test, feature = "test_utils"))]
+#[cfg(feature = "test_utils")]
 pub mod test_util;
 
 #[cfg(test)]
+#[cfg(feature = "test_utils")]
 mod test;
 
 pub mod fixt;