+++ crates/holochain/CHANGELOG.md
@@ -3,6 +3,11 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Refactor conductor to use parking lot rw lock instead of tokio rw lock. (Faster and prevents deadlocks.)
+
+### Changed
+
+- The scheduler should work now
 
 - Batch peer store write so we use less transactions. [#1007](https://github.com/holochain/holochain/pull/1007/).